refactor: useCardType을 getCardType 유틸 함수로 변경
